+++ scripts/create_package.py
@@ -6,7 +6,7 @@ from pathlib import Path
 def main():
     ev3_dir = Path(__file__).resolve().parent.parent
     os.chdir(ev3_dir)
-    subprocess.call(['python', 'setup.py', 'bdist_wheel', '-d', 'pypi'])
+    subprocess.call(['python3', 'setup.py', 'bdist_wheel', '-d', 'pypi'])
 
 
 if __name__ == '__main__':